apps/api/requirements.txt
aenum==3.1.17
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.87.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudevents==1.12.0
cryptography==46.0.6
Deprecated==1.3.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.2
Flask==3.1.3
frozenlist==1.8.0
functions-framework==3.10.1
future==1.0.0
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
google-cloud-core==2.5.1
google-cloud-firestore==2.26.0
google-genai==1.69.0
googleapis-common-protos==1.73.1
grpcio==1.80.0
grpcio-status==1.80.0
gspread==6.2.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
line-bot-sdk==3.22.0
MarkupSafe==3.0.3
multidict==6.7.1
oauthlib==3.3.1
packaging==26.0
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.42.0
uvicorn-worker==0.4.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.7
wrapt==2.1.2
yarl==1.23.0
